+++ lindera/src/token_filter/japanese_base_form.rs
@@ -145,7 +145,7 @@ mod tests {
 
     #[cfg(feature = "ipadic")]
     #[test]
-    fn test_japanese_base_form_token_filter_apply() {
+    fn test_japanese_base_form_token_filter_apply_ipadic() {
         let config_str = r#"
         {
             "kind": "ipadic"
@@ -207,7 +207,7 @@ mod tests {
                     "基本形".to_string(),
                     "ます".to_string(),
                     "マス".to_string(),
-                    "マスS".to_string(),
+                    "マス".to_string(),
                 ]),
             },
         ];
@@ -219,5 +219,138 @@ mod tests {
         assert_eq!(tokens[1].text, "に");
         assert_eq!(tokens[2].text, "ある");
         assert_eq!(tokens[3].text, "ます");
+    }
+
+    #[cfg(feature = "unidic")]
+    #[test]
+    fn test_japanese_base_form_token_filter_apply_unidic() {
+        let config_str = r#"
+        {
+            "kind": "unidic"
+        }
+        "#;
+        let filter = JapaneseBaseFormTokenFilter::from_slice(config_str.as_bytes()).unwrap();
+
+        let mut tokens: Vec<Token> = vec![
+            Token {
+                text: Cow::Borrowed("羽田"),
+                details: Some(vec![
+                    "名詞".to_string(),
+                    "固有名詞".to_string(),
+                    "人名".to_string(),
+                    "姓".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "ハタ".to_string(),
+                    "ハタ".to_string(),
+                    "羽田".to_string(),
+                    "ハタ".to_string(),
+                    "羽田".to_string(),
+                    "ハタ".to_string(),
+                    "固".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                ]),
+            },
+            Token {
+                text: Cow::Borrowed("空港"),
+                details: Some(vec![
+                    "名詞".to_string(),
+                    "普通名詞".to_string(),
+                    "一般".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "クウコウ".to_string(),
+                    "空港".to_string(),
+                    "空港".to_string(),
+                    "クーコー".to_string(),
+                    "空港".to_string(),
+                    "クーコー".to_string(),
+                    "漢".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                ]),
+            },
+            Token {
+                text: Cow::Borrowed("に"),
+                details: Some(vec![
+                    "助詞".to_string(),
+                    "格助詞".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "ニ".to_string(),
+                    "に".to_string(),
+                    "に".to_string(),
+                    "ニ".to_string(),
+                    "に".to_string(),
+                    "ニ".to_string(),
+                    "和".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                ]),
+            },
+            Token {
+                text: Cow::Borrowed("あり"),
+                details: Some(vec![
+                    "動詞".to_string(),
+                    "非自立可能".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "五段-ラ行".to_string(),
+                    "連用形-一般".to_string(),
+                    "アル".to_string(),
+                    "有る".to_string(),
+                    "あり".to_string(),
+                    "アリ".to_string(),
+                    "ある".to_string(),
+                    "アル".to_string(),
+                    "和".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                ]),
+            },
+            Token {
+                text: Cow::Borrowed("ます"),
+                details: Some(vec![
+                    "助動詞".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "助動詞-マス".to_string(),
+                    "終止形-一般".to_string(),
+                    "マス".to_string(),
+                    "ます".to_string(),
+                    "ます".to_string(),
+                    "マス".to_string(),
+                    "ます".to_string(),
+                    "マス".to_string(),
+                    "和".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                    "*".to_string(),
+                ]),
+            },
+        ];
+
+        filter.apply(&mut tokens).unwrap();
+
+        assert_eq!(tokens.len(), 5);
+        assert_eq!(tokens[0].text, "羽田");
+        assert_eq!(tokens[1].text, "空港");
+        assert_eq!(tokens[2].text, "に");
+        assert_eq!(tokens[3].text, "ある");
+        assert_eq!(tokens[4].text, "ます");
     }
 }